create roomController.js and move room route into itfrom route/index.js
