+++ routes/index.js
@@ -1,15 +1,8 @@
 const express = require('express');
-const mongoose = require('mongoose');
-const Room = mongoose.model('Room');
 const router = express.Router();
+const { catchErrors } = require('../handlers/errorHandlers');
 
-catchErrors = (fn) => {
-  return function(req, res, next) {
-    return fn(req, res, next).catch(next);
-  }
-}
-
-//rap catchError for each route
+const roomController = require('../controllers/roomController');
 
 router.get('/', (req, res) => {
   res.render("index");
@@ -20,16 +13,7 @@ router.get('/room/new', (req, res) => {
   res.render("new");
 });
 
-
-router.post('/room/create', catchErrors(async (req, res) => {
-  const roomName = req.body['room-name'];
-  const room = new Room({'room_name': roomName});
-  await room.save();
-
-
-  console.log('it worked!');
-  res.redirect('/');
-}));
+router.post('/room/create', catchErrors(roomController.createRoom));
 
 router.get('/room/edit', (req, res) =>{
   res.render("edit");
